+++ painel-adm/produtos/excluir.php
@@ -9,7 +9,10 @@ $id = $_POST['id'];
 $query_con = $pdo->query("SELECT * FROM produtos WHERE id = '$id'");
 $res_con = $query_con->fetchAll(PDO::FETCH_ASSOC);
 $imagem = $res_con[0]['foto'];
+
+if($imagem != 'sem-foto.jpg') {
 unlink('../../img/produtos/'.$imagem);
+}
 
 //EXCLUSÃO DA CATEGORIA
 $query_verif = $pdo->query("DELETE FROM produtos WHERE id = '$id'"); 
@@ -19,5 +22,3 @@ os dados não estão vindo de um formulário
 */
 
 echo "Excluído com Sucesso!";
-
-?>
